Rename _misc.py to _diag.py
