+++ scico/linop/_diag.py
