+++ scico/linop/_matrix.py
@@ -24,8 +24,8 @@ from jax.interpreters.xla import DeviceArray
 import scico.numpy as snp
 from scico.typing import JaxArray
 
+from ._diag import Identity
 from ._linop import LinearOperator
-from ._misc import Identity
 
 
 def _wrap_add_sub_matrix(func, op):
